# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Attendance Software API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get an employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /employees/{id}/face:
    post:
      operationId: registerFace
      tags: [employees]
      summary: Register or update face descriptors for an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaceDescriptorInput"
      responses:
        "200":
          description: Face registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: removeFace
      tags: [employees]
      summary: Remove face descriptors for an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Face removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"

  # ─── Locations ──────────────────────────────────────────────────────────────
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List all office locations
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [locations]
      summary: Create a new office location with geofence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /locations/{id}:
    get:
      operationId: getLocation
      tags: [locations]
      summary: Get a location by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Location found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLocation
      tags: [locations]
      summary: Update a location (including resizing the geofence radius)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
    delete:
      operationId: deleteLocation
      tags: [locations]
      summary: Delete a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Attendance ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth
...
    Employee:
      type: object
      required: [id, name, email, role, isActive, createdAt, hasFaceRegistered]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        employeeCode:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, employee]
        isActive:
          type: boolean
        hasFaceRegistered:
          type: boolean
        createdAt:
          type: string
          format: date-time
    EmployeeInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        employeeCode:
          type: string
        department:
          type: string
        role:
          type: string
          enum: [admin, employee]
    EmployeeUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        employeeCode:
          type: string
        department:
          type: string
        role:
          type: string
          enum: [admin, employee]
        isActive:
          type: boolean
        password:
          type: string
    FaceDescriptorInput:
      type: object
      required: [descriptors]
      properties:
        descriptors:
          type: array
          description: Array of face descriptor arrays (each 128-element Float32Array serialized as number[])
          items:
            type: array
            items:
              type: number

    # Locations
    Location:
      type: object
      required: [id, name, latitude, longitude, radius, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        radius:
          type: number
          description: Geofence radius in meters
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    LocationInput:
      type: object
      required: [name, latitude, longitude, radius]
      properties:
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        radius:
          type: number
          description: Geofence radius in meters (default 100)
    LocationUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        radius:
          type: number
          description: Geofence radius in meters
        isActive:
          type: boolean

    # Attendance
...