# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Attendance Software API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeRole } from './employeeRole';

export interface Employee {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  employeeCode?: string | null;
  /** @nullable */
  department?: string | null;
  role: EmployeeRole;
  isActive: boolean;
  hasFaceRegistered: boolean;
  createdAt: Date;
}