# lib/api-zod/src/generated/types/employeeInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Attendance Software API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeInputRole } from './employeeInputRole';

export interface EmployeeInput {
  name: string;
  email: string;
  password: string;
  employeeCode?: string;
  department?: string;
  role: EmployeeInputRole;
}